A more finished version of the game
After 3 months and noticing some major errors, I've decided to fix them.
I've removed the colorama module, because apperantly it's a "violation of privacy" for some people (thanks reddit).
Also fixed the clear function, witch up untill now worked only on Windows. I think I made it so that it works on both Windows and Linux, but at the time of this commit, it has only been tested on Linux. If it doesn't work on Windows, I'll fix it in a future commit.